check if MSI is active or use the inventory stock.
