+++ etc/module.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Aplazo_AplazoPayment" setup_version="3.2.5">
+    <module name="Aplazo_AplazoPayment" setup_version="3.2.6">
         <sequence>
             <module name="Magento_Sales"/>
             <module name="Magento_Payment"/>
